+++ SearchBoxes.php
@@ -39,7 +39,7 @@ $wgExtensionCredits['parserhook'][] = array(
 	'author' => array( 'Dror S. [FFS] ([http://www.kolzchut.org.il Kol-Zchut])' ),
 	//'url'            => '',
 	'descriptionmsg' => 'searchboxes-desc',
-	'version'        => '0.1.0',
+	'version'        => '0.2.0',
 );
 
 // Internationalization