+++ CHANGELOG.md
@@ -4,6 +4,12 @@ Revision history for Extension:SearchBoxes
 All notable changes to this project will be documented in this file.
 This project adheres (or attempts to adhere) to [Semantic Versioning](http://semver.org/).
 
+## [0.2.0] - 2015-08-20
+* New options: 'fancybutton', 'elementsize', 'inline', 'hiddenbuttonlabel'
+* remove option 'size' in favor of the new 'widthclasses', which can use any of bootstrap's grid
+  options: col-(xs|sm|md|lg)-1..12.
+* Remove unused option 'break'
+
 ## [0.1.0] - 2015-08-20
 
 Initial release. Two types of search forms: